+++ src/com.tutorial.main/HUD.java
@@ -3,15 +3,15 @@ package com.tutorial.main;
 import java.awt.*;
 
 public class HUD {
-    public static int HEALTH = 100;
-    private int greenVal=255;
+    public static float HEALTH = 100;
+    private float greenVal=255;
 
     private int score =0;
     private int level =1;
 
     public void tick(){
-        HEALTH = Game.clamp(HEALTH, 0, 100);
-        greenVal = Game.clamp(greenVal, 0,255 );
+        HEALTH = Game.clamp((int)HEALTH, 0, 100);
+        greenVal = Game.clamp((int)greenVal, 0,255 );
         greenVal= HEALTH*2;
 
         score++;
@@ -20,8 +20,8 @@ public class HUD {
     public void render(Graphics g){
         g.setColor(Color.gray);
         g.fillRect(15,15,200,32);
-        g.setColor(new Color(75, greenVal, 0 ));
-        g.fillRect(15,15,HEALTH*2,32);
+        g.setColor(new Color(75, (int)greenVal, 0 ));
+        g.fillRect(15,15,(int)HEALTH*2,32);
         g.setColor(Color.white);
         g.drawRect(15,15,200,32);
